Fusion D-D update
made it so that you can switch between D-D and D-T in flight. Just need
to figure out why the max thrust is incorrect. Also, it takes acouple of
clicks on the change mode for it to start working.... after things start
running smoothly I'll add in some of the features that I dissabled like
setting the isp.